Tracks historical IDs
